+++ test/integration/open_api/events_test.rb
@@ -12,6 +12,21 @@ class OpenApi::EventsTest < ActionDispatch::IntegrationTest
   test 'list all events' do
     get '/open_api/v1/events', headers: open_api_headers(@token)
     assert_response :success
+    events = json_response(response.body)
+    assert_not_empty events[:events]
+
+    assert(events[:events].all? { |event| !event[:id].nil? })
+    assert(events[:events].all? { |event| !event[:title].nil? })
+    assert(events[:events].all? { |event| !event[:description].nil? })
+    assert(events[:events].all? { |event| !event[:updated_at].nil? })
+    assert(events[:events].all? { |event| !event[:created_at].nil? })
+    assert(events[:events].all? { |event| !event[:nb_total_places].nil? })
+    assert(events[:events].all? { |event| !event[:nb_free_places].nil? })
+    assert(events[:events].all? { |event| !event[:start_at].nil? })
+    assert(events[:events].all? { |event| !event[:end_at].nil? })
+    assert(events[:events].all? { |event| !event[:category].nil? })
+    assert(events[:events].all? { |event| !event[:prices][:normal].nil? })
+    assert(events[:events].all? { |event| !event[:url].nil? })
   end
 
   test 'list all events with pagination' do